refactor: move MenuItem into global scope
