+++ src/components/HeaderMenu/components/SortSubmenu/index.tsx
@@ -2,7 +2,7 @@ import { SORT_MENU_ITEM } from "../../constants";
 
 import { SortDir } from "src/services/tableState/types";
 
-import MenuItem from "../MenuItem";
+import MenuItem from "../../../MenuItem";
 import Submenu from "../Submenu";
 
 interface Props {
