--- src/components/RowMenu/components/RowMenuItem/styles.css
@@ -1,3 +1,0 @@
-.NLT__drag-menu-item:focus-visible {
-    outline: 1px solid blue;
-}
