+++ src/components/RowMenu/styles.css
@@ -1,4 +1,3 @@
-.NLT__drag-menu {
+.NLT__row-menu {
 	font-size: 0.9rem;
-	padding: 4px 10px;
 }
